# jag_api_contract_v1.yaml
openapi: 3.1.0
info:
  title: JAG Platform API
  version: 1.0.0
  description: |
    JAG Holdings unified platform API — Core, IMS, JABCO PM, CRM, Entertainment, Family, and Properties modules.

    **Authentication:** Keycloak bearer JWT required on all endpoints except `GET /health`.
    Tokens are issued by the Keycloak realm; no custom auth endpoints exist in this API.

    **Response envelope (STD-06):** Every response uses `{ success, data, error, code, meta? }`.
    Errors never expose raw stack traces.

    **Idempotency (STD-11):** All financial write endpoints require `Idempotency-Key: <uuid>`.
    A duplicate key returns the first result without re-execution; response status 409.

    **Versioning (STD-05):** All endpoints are prefixed `/api/v1/`. Breaking changes → `/api/v2/`.

    **Pagination:** All list endpoints accept `?page=&limit=&sort=&order=` and return
    `meta: { total, page, limit, pages }`.

    ---
    ### PRE-2 Design Decisions

    1. **pending_review_queue gap** — flagged in PRE-1, not yet in schema. Failed/disputed WiPay
       webhook events currently return 422 `WIPAY_VALIDATION_FAILED`. If a `pending_review_queue`
       table is added (jag_core or jag_properties), the webhook handler should INSERT a row and
       return 202 instead. Raise in PRE-3/PRE-4.

    2. **entity_tag** — `BAR | MEMBERS_CLUB` is a required field on every Entertainment
       transaction request body and response. Non-negotiable per PRE-1 design decision 4.

...

...

...
servers:
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak-issued JWT. Obtain from the JAG Keycloak realm token endpoint.
...
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      schema:
        type: string
    OrderParam:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  schemas:
    PaginationMeta:
      type: object
      required: [total, page, limit, pages]
      properties:
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        pages:
          type: integer
          example: 3
    ApiResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          description: Response payload — schema varies per endpoint. Null on error.
          nullable: true
        meta:
          $ref: '#/components/schemas/PaginationMeta'
        error:
          type: string
          nullable: true
          description: Human-readable error message. Present only on failure.
        code:
          type: string
          nullable: true
          description: >
            Machine-readable error code. Present only on failure.
            Examples: VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND,
            IDEMPOTENCY_CONFLICT, BUSINESS_RULE_VIOLATION, INTERNAL_ERROR.

    # -------------------------------------------------------
    # CORE SCHEMAS
    # -------------------------------------------------------
...
    CrmCompany:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        name:
          type: string
        industry:
          type: string
          nullable: true
        country:
          type: string
          example: TT
        phone:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        website:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CrmCompanyCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        industry:
          type: string
          nullable: true
        country:
          type: string
          default: TT
        phone:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        website:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
    CrmContact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
          nullable: true
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
          example: Project Manager
        preferred_language:
          type: string
          enum: [en, zh, es]
        loyalty_member_id:
          type: string
          format: uuid
          nullable: true
          description: Cross-DB logical ref to jag_family.fam_loyalty_programmes
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CrmContactCreate:
      type: object
      required: [first_name, last_name]
      properties:
        company_id:
          type: string
          format: uuid
          nullable: true
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
        preferred_language:
          type: string
          enum: [en, zh, es]
          default: en
        notes:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      required: [success, data, error, code]
      properties:
        success:
          type: boolean
          enum: [false]
        data:
          nullable: true
          default: null
        error:
          type: string
          description: Human-readable error message.
        code:
          type: string
          description: Machine-readable error code.
  responses:
    BadRequest:
      description: Validation error (STD-10 — Zod validation failure)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            data: null
            error: "Field 'email' must be a valid email address."
            code: VALIDATION_ERROR
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            data: null
            error: Authentication required.
            code: UNAUTHORIZED
...
paths:
...
  /crm/companies:
    get:
      tags: [CRM]
      summary: List companies
      operationId: listCrmCompanies
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/OrderParam'
      responses:
        '200':
          description: Company list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/CrmCompany'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [CRM]
      summary: Create company
      operationId: createCrmCompany
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrmCompanyCreate'
      responses:
        '201':
          description: Company created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CrmCompany'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /crm/contacts:
    get:
      tags: [CRM]
      summary: List contacts
      operationId: listCrmContacts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/OrderParam'
        - name: company_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Contact list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/CrmContact'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [CRM]
      summary: Create contact
      operationId: createCrmContact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrmContactCreate'
      responses:
        '201':
          description: Contact created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CrmContact'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...